test: update additional tensor dtypes to double in `test_jit_example`

+++ tests/test_jit_example.py
@@ -65,7 +65,7 @@ struct SingleDecodeWithCustomMask {
         ["custom_mask"],  # additional_tensor_names
         ["uint8_t"],  # additional_tensor_dtypes
         ["sm_scale"],  # # additional_scalar_names
-        ["float"],  # additional_scalar_dtypes
+        ["double"],  # additional_scalar_dtypes
         "SingleDecodeWithCustomMask",
         variant_decl,
     )
@@ -167,7 +167,7 @@ def test_flash_sigmoid():
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["logits_scale", "sigmoid_bias"],  # additional_scalar_names
-        ["float", "float"],  # additional_scalar_dtypes
+        ["double", "double"],  # additional_scalar_dtypes
         "FlashSigmoid",
         variant_decl,
     )
@@ -239,7 +239,7 @@ struct DumpLogits {
         ["output_logits"],  # additional_tensor_names
         ["float"],  # additional_tensor_dtypes
         ["sm_scale"],  # additional_scalar_names
-        ["float"],  # additional_scalar_dtypes
+        ["double"],  # additional_scalar_dtypes
         "DumpLogits",
         variant_decl,
     )
@@ -274,7 +274,7 @@ def test_batch_decode_flash_sigmoid(use_tensor_cores):
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["logits_scale", "sigmoid_bias"],  # additional_scalar_names
-        ["float", "float"],  # additional_scalar_dtypes
+        ["double", "double"],  # additional_scalar_dtypes
         "FlashSigmoid",
         variant_decl,
     )
@@ -381,7 +381,7 @@ def test_batch_prefill_flash_sigmoid():
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["logits_scale", "sigmoid_bias"],  # additional_scalar_names
-        ["float", "float"],  # additional_scalar_dtypes
+        ["double", "double"],  # additional_scalar_dtypes
         "FlashSigmoid",
         variant_decl,
     )
@@ -500,7 +500,7 @@ def test_batch_prefill_sm90_flash_sigmoid():
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["logits_scale", "sigmoid_bias"],  # additional_scalar_names
-        ["float", "float"],  # additional_scalar_dtypes
+        ["double", "double"],  # additional_scalar_dtypes
         "FlashSigmoid",
         variant_decl,
     )
@@ -656,7 +656,7 @@ struct DebugPrintLogits {
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["sm_scale"],  # additional_scalar_names
-        ["float"],  # additional_scalar_dtypes
+        ["double"],  # additional_scalar_dtypes
         "DebugPrintLogits",
         variant_decl,
     )
@@ -733,7 +733,7 @@ struct DebugPrintLogits {
         [],  # additional_tensor_names
         [],  # additional_tensor_dtypes
         ["sm_scale"],  # additional_scalar_names
-        ["float"],  # additional_scalar_dtypes
+        ["double"],  # additional_scalar_dtypes
         "DebugPrintLogits",
         variant_decl,
     )
